project/requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
Django==6.0
django-browser-reload==1.21.0
django-filter==25.2
django-tailwind==4.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==40.1.0
gunicorn==23.0.0
honcho==2.0.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg-pool==3.3.0
Pygments==2.19.2
PyJWT==2.10.1
pytailwindcss==0.3.0
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
stripe==13.0.1
rich==14.2.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
